requirements.txt
cachetools
einops~=0.8.0
ipdb
nltk~=3.9.1
omegaconf~=2.4.0.dev3
packaging~=24.2
pytorch_lightning
PyYAML~=6.0.2
tqdm~=4.67.0
SentencePiece
transformers~=4.46.2
deepspeed
torch~=2.5.1
peft~=0.13.2
numpy~=2.1.3
tensorboard
accelerate
transformers
pytorch-lightning~=2.4.0
requests~=2.32.3
pillow~=11.0.0